// src/types/FabloConfigExtended.ts
export interface OrdererConfig {
  name: string;
  domain: string;
  address: string;
  adminPort: number;
  port: number;
  fullAddress: string;
  orgName: string;
  orgMspName: string;
}

export interface CAConfig {
  address: string;
  caAdminNameVar: string;
  caAdminPassVar: string;
  exposePort: number;
  fullAddress: string;
  port: number;
  prefix: string;
  db: "sqlite" | "postgres" | "mysql";
}

export interface PeerConfig {
  address: string;
  couchDbExposePort: number;
  db: PeerDbConfig;
  fullAddress: string;
  isAnchorPeer: boolean;
  name: string;
  port: number;
  gatewayEnabled: boolean;
}

export interface PeerDbConfig {
  type: "LevelDb" | "CouchDb";
  image?: string;
}

export interface CLIConfig {
  address: string;
}
// ...
export interface FabloRestLoggingConfig {
  info?: "console" | string;
  warn?: "console" | string;
  error?: "console" | string;
  debug?: "console" | string;
}

export interface FabloRestConfig {
  address: string;
  port: number;
  mspId: string;
  fabricCaUrl: string;
  fabricCaName: string;
  discoveryUrls: string;
  discoverySslTargetNameOverrides: string;
  discoveryTlsCaCertFiles: string;
  logging: FabloRestLoggingConfig;
}

export interface ExplorerConfig {
  address: string;
  port: number;
}

export interface OrgConfig {
  anchorPeers: PeerConfig[];
  bootstrapPeers: string;
  ca: CAConfig;
  cli: CLIConfig;
  cryptoConfigFileName: string;
  domain: string;
  headPeer?: PeerConfig;
  mspName: string;
  name: string;
  peers: PeerConfig[];
  peersCount: number;
  ordererGroups: OrdererGroup[];
  tools: { fabloRest?: FabloRestConfig; explorer?: ExplorerConfig };
}
// ...
export interface OrdererGroup {
  name: string;
  consensus: "solo" | "etcdraft" | "BFT";
  profileName: string;
  genesisBlockName: string;
  configtxOrdererDefaults: string;
  hostingOrgs: string[];
  orderers: OrdererConfig[];
  ordererHeads: OrdererConfig[];
}

export interface HooksConfig {
  postGenerate: string;
  postStart: string;
}
export interface NamespaceConfig {
  name: string;
  policy: string;
}